+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - TimePicker component by [@amit-y](https://github.com/amit-y)
+- DateTimePicker component by [@amit-y](https://github.com/amit-y)
 
 ## [1.17.0] - 2023-12-14